requirements.txt
Flask==1.0.3
Flask-SocketIO==4.1.0
gunicorn==19.9.0
requests
flask_session
boto3
python-decouple
eventlet
werkzeug
